refactor(format): move SplitInfo to separate file.
